Fix compressed headers check in backfill

+++ nntmux/Backfill.php
@@ -111,7 +111,7 @@ class Backfill
 			}
 		}
 
-		$this->_compressedHeaders = Settings::value('..compressedheaders') === 1 ? true : false;
+		$this->_compressedHeaders = (int)Settings::value('..compressedheaders') === 1 ? true : false;
 		$this->_safeBackFillDate = Settings::value('..safebackfilldate') !== '' ? (string)Settings::value('safebackfilldate') : '2008-08-14';
 		$this->_safePartRepair = Settings::value('..safepartrepair') === 1 ? 'update' : 'backfill';
 		$this->_disableBackfillGroup = Settings::value('..disablebackfillgroup') === 1 ? true : false;
